@@ -4,48 +4,46 @@
   "license": "Apache-2.0",
   "author": "Zooniverse <contact@zooniverse.org> (https://www.zooniverse.org/)",
   "version": "0.0.1",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
+  "type": "module",
+  "main": "dist/esm/index.js",
   "exports": {
     ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
+      "import": "./dist/esm/index.js"
     },
     "./*": {
-      "import": "./dist/esm/*/index.js",
-      "require": "./dist/cjs/*/index.js"
+      "import": "./dist/esm/*/index.js"
     }
   },
   "sideEffects": false,
   "repository": "https://github.com/zooniverse/front-end-monorepo.git",
   "bugs": "https://github.com/zooniverse/front-end-monorepo/issues",
   "scripts": {
-    "build": "yarn build:cjs && yarn build:es6",
-    "build:cjs": "babel ./src/ --out-dir ./dist/cjs --copy-files --no-copy-ignored --ignore 'src/**/*.spec.js' --ignore 'src/**/*.stories.js'",
-    "build:es6": "BABEL_ENV=es6 babel ./src/ --out-dir ./dist/esm --copy-files --no-copy-ignored --ignore 'src/**/*.spec.js' --ignore 'src/**/*.stories.js'",
+    "build": "yarn build:es6",
+    "build:es6": "BABEL_ENV=es6 babel ./src/ --out-dir ./dist/esm --copy-files --no-copy-ignored --ignore 'src/**/*.spec.jsx' --ignore 'src/**/*.stories.jsx'",
     "prepare": "yarn build",
-    "test": "mocha --config ./test/.mocharc.json ./.storybook/specConfig.js \"./src/**/*.spec.js\"",
-    "test:ci": "mocha --config ./test/.mocharc.json ./.storybook/specConfig.js --reporter=min \"./src/**/*.spec.js\"",
+    "test": "vitest run",
+    "test:ci": "vitest run",
+    "coverage": "vitest run --coverage",
     "storybook": "storybook dev -p 6005",
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@next/third-parties": "~14.2.26",
-    "@zooniverse/async-states": "~0.0.1",
-    "@zooniverse/react-components": "~1.14.0",
     "dayjs": "~1.11.11",
     "i18next": "~24.2.2",
     "panoptes-client": "~5.6.0",
     "prop-types": "~15.8.1",
-    "react-i18next": "~14.1.3"
+    "react-i18next": "~14.1.3",
+    "swr": "~2.3.2"
   },
   "peerDependencies": {
+    "@next/third-parties": "~14.2.26",
     "@zooniverse/grommet-theme": "3.x.x",
+    "@zooniverse/react-components": "~1.14.0",
     "grommet": "2.x.x",
     "grommet-icons": "4.x.x",
     "next": "~14.2.26",
-    "react": "^17.0.0 || ^18.0.0",
-    "react-dom": "^17.0.0 || ^18.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "styled-components": "6.x.x"
   },
   "devDependencies": {
@@ -59,18 +57,15 @@
     "@testing-library/dom": "~9.3.0",
     "@testing-library/react": "~14.2.0",
     "@testing-library/user-event": "~14.5.0",
+    "@vitejs/plugin-react": "~4.4.1",
+    "babel-plugin-module-resolver": "~5.0.0",
     "babel-plugin-styled-components": "~2.1.4",
     "babel-plugin-transform-imports": "~2.0.0",
-    "chai": "~4.5.0",
-    "chai-dom": "~1.12.0",
-    "dirty-chai": "~2.0.1",
     "eslint-plugin-jsx-a11y": "~6.10.1",
     "jsdom": "~26.1.0",
-    "mocha": "~11.4.0",
-    "sinon": "~20.0.0",
-    "sinon-chai": "~3.7.0",
     "storybook": "~7.6.11",
-    "storybook-react-i18next": "~2.0.1"
+    "storybook-react-i18next": "~2.0.1",
+    "vitest": "~3.1.3"
   },
   "engines": {
     "node": ">=20.5 <=20.18"
