@@ -1,5 +1,7 @@
+import vitetestConfig from '../vitest.config.js'
+
 const config = {
-  stories: ['../src/**/*.stories.js'],
+  stories: ['../src/**/*.stories.jsx'],
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-essentials',
@@ -11,6 +13,18 @@ const config = {
   },
   docs: {
     autodocs: 'tag'
+  },
+  webpackFinal: async config => {
+    return {
+      ...config,
+      resolve: {
+        ...vitetestConfig.resolve,
+        extensions: ['.jsx', '.js', '...'],
+        fallback: {
+          fs: false
+        }
+      }
+    }
   }
 }
 export default config
