@@ -1,5 +1,7 @@
+import vitetestConfig from '../vitest.config.js'
+
 const config = {
-  stories: ['../src/**/*.stories.js'],
+  stories: ['../src/**/*.stories.jsx'],
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-essentials'
@@ -12,5 +14,17 @@ const config = {
     autodocs: 'tag'
   },
   staticDirs: ['../storybook-public'], // needed for msw-storybook-addon
+    webpackFinal: async config => {
+    return {
+      ...config,
+      resolve: {
+        ...vitetestConfig.resolve,
+        extensions: ['.jsx', '.js', '...'],
+        fallback: {
+          fs: false
+        }
+      }
+    }
+  }
 }
 export default config
