@@ -1,1 +1,1 @@
-export { default } from './BarChart.js'
+export { default } from './BarChart'
