@@ -1,1 +1,1 @@
-export { default } from './Pagination.js'
+export { default } from './Pagination'
