@@ -1,1 +1,1 @@
-export { default } from './Select.js'
+export { default } from './Select'
